Add files via upload
I have considered a sample dataset of Vrinda store which depicts the annual sales report of the company. I have used MICROSOFT EXCEL in order to describe a following set of necessary analytical fields that I will list below
The questions that I have analysed in the excel project are as follows-
1)Which month got the highest sales and orders?
2)Who purchased more-men or women in 2022?
3)What are different order status in 2022?
4)List top 5 states contributing to the sales?
5)Relation between age and gender based on number of orders
6)Which channel is contributing to maximum sales?

there are two files attached, one is the excel tabular form that I have obtained from Internet and the other file is the my analysis file.